+++ eslint.config.mjs
@@ -12,6 +12,7 @@ export default tseslint.config(
     languageOptions: {
       globals: {
         console: 'readonly',
+        process: 'readonly',
       },
     },
   },